Adding picture upload option to recipes. Pending stability.

+++ server/routers/recipeRouter.js
@@ -1,4 +1,6 @@
 const express = require('express');
+const multer = require('multer');
+const sharp = require('sharp');
 const Recipe = require('../models/recipeModel.js');
 const User = require('../models/userModel.js');
 const auth = require('../middleware/authentication.js');
@@ -33,7 +35,7 @@ router.get('/recipes/all', async (req, res) => {
 
 //Get a user's submitted recipes.
 //NEEDS PAGNATION
-router.get('/recipes/:username', async (req, res) => {
+router.get('/recipes/user/:username', async (req, res) => {
     try {
         const user = await User.findOne({ username: req.params.username });
         const recipes = await Recipe.find({ owner: user._id });
@@ -90,7 +92,7 @@ router.patch('/recipes/:id', auth, async (req, res) => {
 //Delete a recipe.
 router.delete('/recipes/:id', auth, async (req, res) => {
     try {
-        const recipe = await Recipe.findOneAndDelete({ _id: req.params.id });
+        const recipe = await Recipe.findOneAndDelete({ _id: req.params.id, owner: req.user._id });
         
         if (!recipe) {
             return res.status(404).send("Cannot find recipe.");
@@ -102,6 +104,82 @@ router.delete('/recipes/:id', auth, async (req, res) => {
     }
 });
 
+
+
+
+//Recipe pictures
+//Documentation - https://www.npmjs.com/package/multer 
+const upload = multer({
+    limits: {
+        filesize: 1000000
+    },
+    fileFilter(req, file, callback) {
+        if (!file.originalname.match(/\.(png|jpg|jpeg|bmp)$/)) {
+            return callback(new Error("Unsupported image file type."));
+        }
+        //If file is good, null, true to accept file.
+        callback(undefined, true);
+    }
+});
+
+//Upload picture(s) of the food
+router.post('/recipes/:id/pictures', auth, upload.array('pictures', 5), async (req, res) => {
+    const recipe = await Recipe.findOne({ _id: req.params.id, owner: req.user._id });
+
+    //const recipe = await Recipe.findOne(req.user._id);
+    // if (req.user._id.toString() !== recipe.owner.toString()) {
+    //     return res.status(403).send("User is not owner of the recipe.");
+    // }
+
+    if (!recipe) {
+        return res.status(404).send("Cannot find recipe.");
+    }
+
+    for (var i = 0; i < req.files.length; i++) {
+        const buffer = await sharp(req.files[i].buffer).resize({ width: 640, height: 360 }).jpeg().toBuffer();
+        recipe.pictures = recipe.pictures.concat({ picture: buffer });
+    }
+
+    await recipe.save();
+    res.send();
+}, (error, req, res, next) => {
+    res.send(400).send(error.message);
+});
+
+//Get picture(s)
+//What if we used pagnation or populate for the individual images in the pictures array? Query strings?
+router.get('/recipes/:id/pictures', async (req, res) => {
+    try {
+        const recipe = await Recipe.findById(req.params.id);
+
+        if (!recipe || !recipe.pictures) {
+            throw new Error("None found.");
+        }
+
+        res.set('Content-Type', 'image/jpeg');
+        res.send(recipe.pictures);
+    } catch (error) {
+        res.status(404).send();
+    }
+});
+
+//Delete picture(s) of the food.
+//What if we used pagnation or populate for the individual images in the pictures array? Query strings?
+router.delete('/recipes/:id/pictures', auth, async (req, res) => {
+    try {
+        const recipe = await Recipe.findOne({ _id: req.params.id, owner: req.user._id });
+
+        if (!recipe || !recipe.pictures) {
+            throw new Error();
+        }
+
+        recipe.pictures = [];
+        await recipe.save();
+        res.send();
+    } catch (error) {
+        res.status(500).send();
+    }
+});
